Remove workaround for old bug
I previously had an issue where the IR receiver seemed to give random output when given the same IR signal. This seemed to start occurring after the Arduino had been on for a certain period of time.
I believe this was caused by faults/inconsistencies in the power supplied to the Arduino, as the receiver started working consistently again after changing a faulty fuse.